# Generate HTML episode description only when needed.