--- a/share/gpodder/extensions/youtube-dl.py
+++ b/share/gpodder/extensions/youtube-dl.py
@@ -211,8 +211,6 @@ class YoutubeFeed(model.Feed):
         episodes = []
         for en in self._ie_result['entries']:
             guid = video_guid(en['id'])
-            description = util.remove_html_tags(en.get('description') or _('No description available'))
-            html_description = util.nice_html_description(en.get('thumbnail'), description)
             if en.get('ext'):
                 mime_type = util.mimetype_from_extension('.{}'.format(en['ext']))
             else:
@@ -225,8 +223,9 @@ class YoutubeFeed(model.Feed):
             ep = {
                 'title': en.get('title', guid),
                 'link': en.get('webpage_url'),
-                'description': description,
-                'description_html': html_description,
+                'episode_art_url': en.get('thumbnail'),
+                'description': util.remove_html_tags(en.get('description') or ''),
+                'description_html': '',
                 'url': en.get('webpage_url'),
                 'file_size': filesize,
                 'mime_type': mime_type,