@elelay 2022-03-16
I like the episode art url. gpodder-core has also added it, downloading the image when downloading the episode, deleting it when deleting the episode. They use it for the embedded player, though: useless to us. Nevermind: not the subject of this PR.
But, please change the column to `episode_art_url` for [consistency](https://github.com/gpodder/gpodder-core/blob/master/src/gpodder/model.py#L97).

I used to be against overriding the description when the author has taken the time to make it different than the html description. But I can't find examples anymore so I'm willing to let it go.

Is performance OK when searching now that the text_description is generated for each keypress (and also for displaying when *display description* is enabled) ? If OK, no objection.
If not, is it possible to have text_description memoized in a transient field?

As there is a schema change, shall we also add the [chapters](https://github.com/gpodder/gpodder-core/blob/master/src/gpodder/model.py#L94) string field so we can add [podlove chapters support in gpodder-core since 2014](https://github.com/gpodder/gpodder-core/commit/3f21e5c0913dd47d8c24242d7be8d8cd328ee923) later? This could go in a different PR but should be in the same release.

Last missing field from gpodder-core episode model would be the subtitle, but I don't really see the point of it. What do you think.