@elelay 2022-03-21
See branch auouymous-ondemand-html-description for 2 commits for chapters+subtitles in addition to your PR.
Not sure yet about subtitle: maybe add the field but don't put anything in it?