@auouymous 2022-03-22
I pushed your changes, and the subtitle and chapters are working for me on https://feeds.metaebene.me/cre/m4a episodes (with my debug extension).

As for subtitle, `info['subtitle'] = episode._text_description` in the tagging extension could use it as `info['subtitle'] = episode.subtitle or episode._text_description`. We could add a smaller line below title in shownotes for episodes with a subtitle, and query it with EQL. Or it could be added before the description in shownotes just in case a feed uses long subtitles. But we shouldn't add it to the database unless we actually plan to use it.