@elelay 2022-03-23
Let's have a look at subtitles in our subscriptions:
 - https://feeds.fireside.fm/bsdnow/rss has a clean one-line subtitle based on description but hand crafted
 - http://feeds.themoth.org/themothpodcast as a one-line subtitle

more examples:
```
sqlite> select podcast.url, e.cnt from (select podcast_id, count(*) as cnt from episode where subtitle is not null group by podcast_id) e join podcast on e.podcast_id = podcast.id;
http://feeds.themoth.org/themothpodcast|209
https://feeds.fireside.fm/bsdnow/rss|207
https://feeds.simplecast.com/BqbsxVfO|1
http://api.sr.se/api/rss/pod/4023|222
https://podcasts.files.bbci.co.uk/p02pc9qx.rss|222
https://www.thetype.com/typechat/feed/|170
http://podcasts.files.bbci.co.uk/p05t3gr2.rss|56
http://feeds.feedburner.com/podnutzpro?format=xml|30
https://radicalpersonalfinance.libsyn.com/rss|165
https://www.thenakedscientists.com/naked_scientists_podcast.xml|222
https://essay-faz.podigee.io/feed/mp3|113
http://cre.fm/feed/m4a|222
https://www.functionalgeekery.com/feed/mp3/|75
https://www.pipes.digital/feed/1N5Aa3Ng|221
https://www.thejazzsession.com/feed/?category_name=podcast|221
https://libwww.freelibrary.org/podcast/rss.cfm|222
https://feeds.podcastmirror.com/makeshiftstories|222
https://onscript.study/feed/podcast|215
http://www.rtve.es/api/programas/2058/audios.rss|20
https://feeds.megaphone.fm/offtopic|114
http://freakshow.fm/feed/m4a|152
```