@auouymous 2022-03-24
http://feeds.themoth.org/themothpodcast `The Moth features people telling true, engaging, funny, touching and eye-opening stories from their lives.` is one line but quite long and would require at least two lines in a narrow shownotes.

`Terravision ist der erste virtuelle Globus, das          
Original hinter Google Earth und die Vorlage für den Film &amp;quot;The         
Billion Dollar Code&amp;quot;` from http://cre.fm/feed/m4a is not only long but also contains HTML entities. I don't know why, but the text shownotes description properly converts HTML entities and the title displays them raw. I'd guess the entities are allowed since subtitle is used as a short description in itunes.

http://feeds.feedburner.com/se-radio, which has an HTML episode description and episode text description contains the channel description, puts the channel name in subtitle. :)
```
title:  Episode 504: Frank McSherry on Materialize
subtitle:  SE-Radio
```