@elelay 2022-06-18
ok to drop subtitle then. Can you make it in this branch?